4.3 version with customcert, adptable, reengagement and recompletion

+++ admin/tool/dataprivacy/classes/task/process_data_request_task.php
@@ -26,6 +26,7 @@ namespace tool_dataprivacy\task;
 
 use action_link;
 use coding_exception;
+use context_system;
 use core\message\message;
 use core\task\adhoc_task;
 use core_user;
@@ -33,8 +34,6 @@ use moodle_exception;
 use moodle_url;
 use tool_dataprivacy\api;
 use tool_dataprivacy\data_request;
-
-defined('MOODLE_INTERNAL') || die();
 
 /**
  * Class that processes an approved data request and prepares/deletes the user's data.
@@ -81,6 +80,7 @@ class process_data_request_task extends adhoc_task {
 
         // Grab the manager.
         // We set an observer against it to handle failures.
+        $allowfiltering = get_config('tool_dataprivacy', 'allowfiltering');
         $manager = new \core_privacy\manager();
         $manager->set_observer(new \tool_dataprivacy\manager_observer());
 
@@ -93,8 +93,6 @@ class process_data_request_task extends adhoc_task {
         $contextlistcollection = $manager->get_contexts_for_userid($requestpersistent->get('userid'));
 
         mtrace('Fetching approved contextlists from collection');
-        $approvedclcollection = api::get_approved_contextlist_collection_for_collection(
-                $contextlistcollection, $foruser, $request->type);
 
         mtrace('Processing request...');
         $completestatus = api::DATAREQUEST_STATUS_COMPLETE;
@@ -102,6 +100,17 @@ class process_data_request_task extends adhoc_task {
 
         if ($request->type == api::DATAREQUEST_TYPE_EXPORT) {
             // Get the user context.
+            if ($allowfiltering) {
+                // Get the collection of approved_contextlist objects needed for core_privacy data export.
+                $approvedclcollection = api::get_approved_contextlist_collection_for_request($requestpersistent);
+            } else {
+                $approvedclcollection = api::get_approved_contextlist_collection_for_collection(
+                    $contextlistcollection,
+                    $foruser,
+                    $request->type,
+                );
+            }
+
             $usercontext = \context_user::instance($foruser->id, IGNORE_MISSING);
             if (!$usercontext) {
                 mtrace("Request {$requestid} cannot be processed due to a missing user context instance for the user
@@ -127,13 +136,22 @@ class process_data_request_task extends adhoc_task {
             $thing = $fs->create_file_from_pathname($filerecord, $exportedcontent);
             $completestatus = api::DATAREQUEST_STATUS_DOWNLOAD_READY;
         } else if ($request->type == api::DATAREQUEST_TYPE_DELETE) {
-            // Delete the data.
-            $manager = new \core_privacy\manager();
-            $manager->set_observer(new \tool_dataprivacy\manager_observer());
+            // Delete the data for users other than the primary admin, which is rejected.
+            if (is_primary_admin($foruser->id)) {
+                $completestatus = api::DATAREQUEST_STATUS_REJECTED;
+            } else {
+                $approvedclcollection = api::get_approved_contextlist_collection_for_collection(
+                    $contextlistcollection,
+                    $foruser,
+                    $request->type,
+                );
+                $manager = new \core_privacy\manager();
+                $manager->set_observer(new \tool_dataprivacy\manager_observer());
 
-            $manager->delete_data_for_user($approvedclcollection);
-            $completestatus = api::DATAREQUEST_STATUS_DELETED;
-            $deleteuser = !$foruser->deleted;
+                $manager->delete_data_for_user($approvedclcollection);
+                $completestatus = api::DATAREQUEST_STATUS_DELETED;
+                $deleteuser = !$foruser->deleted;
+            }
         }
 
         // When the preparation of the metadata finishes, update the request status to awaiting approval.
@@ -141,14 +159,20 @@ class process_data_request_task extends adhoc_task {
         mtrace('The processing of the user data request has been completed...');
 
         // Create message to notify the user regarding the processing results.
-        $dpo = core_user::get_user($request->dpo);
         $message = new message();
         $message->courseid = $SITE->id;
         $message->component = 'tool_dataprivacy';
         $message->name = 'datarequestprocessingresults';
-        $message->userfrom = $dpo;
-        $message->replyto = $dpo->email;
-        $message->replytoname = fullname($dpo->email);
+        if (empty($request->dpo)) {
+            // Use the no-reply user as the sender if the privacy officer is not set. This is the case for automatically
+            // approved requests.
+            $fromuser = core_user::get_noreply_user();
+        } else {
+            $fromuser = core_user::get_user($request->dpo);
+            $message->replyto = $fromuser->email;
+            $message->replytoname = fullname($fromuser);
+        }
+        $message->userfrom = $fromuser;
 
         $typetext = null;
         // Prepare the context data for the email message body.
@@ -176,7 +200,8 @@ class process_data_request_task extends adhoc_task {
                 $message->contexturl = $datarequestsurl;
                 $message->contexturlname = get_string('datarequests', 'tool_dataprivacy');
                 // Message to the recipient.
-                $messagetextdata['message'] = get_string('resultdownloadready', 'tool_dataprivacy', $SITE->fullname);
+                $messagetextdata['message'] = get_string('resultdownloadready', 'tool_dataprivacy',
+                    format_string($SITE->fullname, true, ['context' => context_system::instance()]));
                 // Prepare download link.
                 $downloadurl = moodle_url::make_pluginfile_url($usercontext->id, 'tool_dataprivacy', 'export', $thing->get_itemid(),
                     $thing->get_filepath(), $thing->get_filename(), true);
@@ -188,7 +213,8 @@ class process_data_request_task extends adhoc_task {
                 // No point notifying a deleted user in Moodle.
                 $message->notification = 0;
                 // Message to the recipient.
-                $messagetextdata['message'] = get_string('resultdeleted', 'tool_dataprivacy', $SITE->fullname);
+                $messagetextdata['message'] = get_string('resultdeleted', 'tool_dataprivacy',
+                    format_string($SITE->fullname, true, ['context' => context_system::instance()]));
                 // Message will be sent to the deleted user via email only.
                 $emailonly = true;
                 break;
@@ -212,7 +238,7 @@ class process_data_request_task extends adhoc_task {
             if ($emailonly) {
                 // Do not sent an email if the user has been deleted. The user email has been previously deleted.
                 if (!$foruser->deleted) {
-                    $messagesent = email_to_user($foruser, $dpo, $subject, $message->fullmessage, $messagehtml);
+                    $messagesent = email_to_user($foruser, $fromuser, $subject, $message->fullmessage, $messagehtml);
                 }
             } else {
                 $messagesent = message_send($message);
@@ -258,7 +284,7 @@ class process_data_request_task extends adhoc_task {
 
                 // Send message.
                 if ($emailonly) {
-                    email_to_user($requestedby, $dpo, $subject, $message->fullmessage, $messagehtml);
+                    email_to_user($requestedby, $fromuser, $subject, $message->fullmessage, $messagehtml);
                 } else {
                     message_send($message);
                 }
